requirements.txt
cffi==1.15.0
cryptography==36.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
